named_entity_recognition/requirements.txt
datasets==2.10.1
pandas==1.5.3
Pillow==9.4.0
pytorch_lightning==2.0.0
scikit_learn==1.2.2
transformers==4.26.1
hydra-colorlog==1.0.0
hydra-core==1.1.1
omegaconf==2.1.0